Added the display.js example script
This script takes an URL representing a cache resource (generally
ignoring the hostname and protocol) and displays the result. If the
result is an image it gets piped to the ImageMagick `display` program
(which must be in the current user's execution path). If the result is
anything other than an image it gets printed to the console.