Added proof-of-concept code for asynchronously logging messages
The mapcache library performs logging via a user specified callback.
The problem is that the callback is called from code running
asynchronously in a thread other than the main Node/V8 event loop.
This means the callback cannot directly interact with the javascript
world and must instead employ an asynchronous mechanism for outputting
the log data to a javascript caller.
<http://nikhilm.github.com/uvbook/threads.html#inter-thread-communication>
describes the problem and solution in more detail.

The current implementation is heavily influenced by
<https://gist.github.com/1247101>.